[BUGFIX beta] Remove link-to keyword.
* `RELATED_VIEW` was actually referencing `this.parentView` so there was
  no need to pass it through specially.
* Change `component:-link-to` registration to `component:link-to`
  (allows users to implement a custom `{{link-to}}` in their apps at
  `app/components/link-to.js` without any special work).
* Move `{{link-to}}` documentation to `ember-routing-htmlbars/components/link-to`.
* `ember-routing-views/views/link` -> `ember-routing-views/components/link-to`
* Move `HAS_BLOCK` symbol to `ComponentNodeManager`.
* Remove `ember-routing-htmlbars/keywords/link-to.js`.

(cherry picked from commit e017e5485bb5fd5ebaeee725c2affe6e919243da)
