--- packages/ember-views/lib/system/link-to.js
@@ -1,5 +1,0 @@
-import { symbol } from 'ember-metal/utils';
-
-// These symbols will be used to limit link-to's public API surface area.
-export let HAS_BLOCK = symbol('HAS_BLOCK');
-export let RELATED_VIEW = symbol('RELATED_VIEW');
